PFM-1050 | Updated contract service | Incorporated works specific changes

+++ backend/contract-service/src/main/java/org/egov/works/repository/rowmapper/AmountBreakupRowMapper.java
@@ -38,7 +38,7 @@ public class AmountBreakupRowMapper implements ResultSetExtractor<List<AmountBre
             String id = rs.getString("id");
             String estimateAmountBreakupId = rs.getString("estimateAmountBreakupId");
             String lineItemId = rs.getString("lineItemId");
-            BigDecimal amount = rs.getBigDecimal("amount");
+            Double amount = rs.getDouble("amount");
             String status = rs.getString("status");
             String createdby = rs.getString("createdby");
             String lastmodifiedby = rs.getString("lastModifiedBy");
