+++ backend/contract-service/src/main/java/org/egov/works/web/models/AmountDetail.java
@@ -23,7 +23,7 @@ public class AmountDetail {
     private String id = null;
 
     @JsonProperty("amount")
-    private BigDecimal amount = null;
+    private Double amount = null;
 
     @JsonProperty("additionalDetails")
     private Object additionalDetails = null;
